# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
        "/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Barcha sevimlilarni olish",
                "operationId": "d4f073b0b3b1fe48795a798e6f32e99f",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Yangi sevimli qo'shish",
                "operationId": "8deb669247d78273637f19d6b9245aec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/favorites/{id}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Sevimlini o‘chirish",
                "operationId": "90c103e7351b370ead31ae2f59698102",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/guides": {
            "get": {
                "tags": [
                    "Guides"
                ],
                "summary": "Barcha guidelarni olish",
                "operationId": "bccb2a59f783da8a2bb55147f02eb492",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Guides"
                ],
                "summary": "Yangi guide yaratish",
                "operationId": "5880e93f568ec9ae97800f0c8309d11e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/guides/{id}": {
            "get": {
                "tags": [
                    "Guides"
                ],
                "summary": "Guideni ko‘rish",
                "operationId": "1a69d97e8824301789a7ae1ef86560be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guides"
                ],
                "summary": "Guideni yangilash",
                "operationId": "1732cd0e2713d10f1206cb5949348c34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guides"
                ],
                "summary": "Guideni o‘chirish",
                "operationId": "4db9e44ad877c99e63a8058611baa2f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/guide-types": {
            "get": {
                "tags": [
                    "GuideTypes"
                ],
                "summary": "Barcha guide turlarini olish",
                "operationId": "8aabd4c95a12b2efc776647178d7ef81",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "GuideTypes"
                ],
                "summary": "Yangi guide turi yaratish",
                "operationId": "511ff03082631134428a6429511ef7c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/guide-types/{id}": {
            "get": {
                "tags": [
                    "GuideTypes"
                ],
                "summary": "Guide turini ko‘rish",
                "operationId": "1428d64cb8e356a3b56535565a1c8eb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "GuideTypes"
                ],
                "summary": "Guide turini yangilash",
                "operationId": "e4c30431fab644014c4c30b4786bbc28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "GuideTypes"
                ],
                "summary": "Guide turini o‘chirish",
                "operationId": "394337a1d0be8eb9bdb9203b8676b282",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/service-categories": {
            "get": {
                "tags": [
                    "ServiceCategories"
                ],
                "summary": "Barcha xizmat kategoriyalarini olish",
                "operationId": "e964fd4617d4e97c4e751d019c5d3f25",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "ServiceCategories"
                ],
                "summary": "Yangi xizmat kategoriyasi yaratish",
                "operationId": "6b45ced7ad652d199461c7dc4c86117b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/service-categories/{id}": {
            "get": {
                "tags": [
                    "ServiceCategories"
                ],
                "summary": "Xizmat kategoriyasini ko‘rish",
                "operationId": "1c74f0f907f6854d92da7ff739adfa43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "ServiceCategories"
                ],
                "summary": "Xizmat kategoriyasini yangilash",
                "operationId": "b6d2e25ff1716caa8db1ea2b2e4b4fa9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ServiceCategories"
                ],
                "summary": "Xizmat kategoriyasini o‘chirish",
                "operationId": "34a21cf0f4d2ea6627ecdbf5232a327b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}